Support using alternate gits configuration files.
Allow $GITSLAVE environmental variable to override the selection of
.gitslave as the configuration file name to use.  This, in theory,
would allow for alternate names for special operations.  For example,
you might want to have a superset of files you tag or branch for
release which include files not directly needed to build.  Another
example is someone working on a subset of the modules (e.g. just the
GUI or just the database) and who do not want to burn the disk to
store (and potentially build) the other bits.