allow gits inits to work without .gitslave again; unbreak --version
Fix botched merge changes that broke help/inits when no .gitslave
present.

Generate version with git commands only at install time.  Uses git
commands to generate tag or hash only if file is unpatched by the
install process; Makefile is modified to get version from local
working copy of gits (assumes that gits is checked out with git)
and patch the copy as it is installed, so that --version works
outside of the gits repository.