Revert Dupuy 78ec87d -- gits branch checkout support
From the description of the problem he had, this was not the correct change
and in any case, it broke golden builds.

The problem seems to be if master has a local branch (HADIT-1) and
you perhaps adjust the GITSLAVE or .gitslave (or failed to create
the local branch using gits), git populate fails because the new
children do not have the HADIT-1 branch.

The fix did not handle this case, it created the HADIT-1 branch on the
top level without the request of the user, but then failed to create
the branch for any of the children.

In any case, gits has no business creating branches without an explicit
request.  golden-build is attempting to check out a tag (a perfectly
valid way to operate) but this change made the top level attempt to create
a local branch from a tag, which really cannot work. .  What I could see being legal would be, if the
top level repo had a local branch and the slave did not, perhaps
automatically creating the local branch for the slave, but even this
is fallable.  Branch and tag operations are really important to create
correctly the first time instead of updating them piecemeal.  Expecting
gits to intuit what the user wanted to have happen is likely to be
very challenging and error-prone.