Allow populate to select ifpresent repositories
You may list ifpresent repositories on the populate command line.
Those repositories will be made present and checked out.

Additionally, if there is a non-git directory at the location of a
ifpresent selection, allow populate to check out to that location
(which makes an alternate way to make something present--just mkdir
the checkout location and then populate).