Allow recursive slaves (with a constant relative directory)
Handle the clone and populate case where the file you include is in a
submodule which you have not checked out yet.  Let it ride the first time
and fail if we don't make progress.

This does not yet fully support generic recursive slaves, though,
since the relative base for the slave's subproject is the
master-superproject, not the sub-superproject.