More precise .gitignore entries
Match the slave path we are adding exactly, instead of anything in the
tree with the same name (file or directory).

Please note that this is an "INCOMPATIBLE" change between older
gitslave versions and this current version.  What this means is that
newer versions of gitslave cannot usefully manage the .gitignore file
when performing a `gits detach` on a superrepo that was created with
an older gitslave `gits attach`.  The reverse (detaching with an older
gitslave that what attached it).  This does not seem especially
serious, but must be documented