add default pagination (like git)
Pagination is only enabled if standard output is a tty (and there is a
controlling tty for the pager to take input from).  If you do not want
paginated output, the -n or --no-pager option disables it.  All the git
pager configuration settings (GIT_PAGER and PAGER environment variables,
and core.pager configuration setting) are honored.

Pagination is on by default unless -v|--verbose is specified (as that
generates output to stderr that messes it up if not redirected)

If -q|--quiet was not specified, and pagination is in effect, redirect
stderr of docmd() git commands to stdout (pager input) for sanity