Add --versionp plus add --porcelain to version.  Document "version" cmd
The "version" subcommand was previously undocumented.  Document it since
we modify the output.

version and --version had different behavior depending on whether the
program was installed or not.  Unify that behavior to the installed
version and add --porcelain flag to version (and gits --versionp
option) to get just the bare gits version for `make install`